@@ -43,26 +43,39 @@
     },
     "require": {
         "php": "^8.0",
-        "doctrine/annotations": "^1.6",
-        "doctrine/doctrine-orm-module": "^6.1",
-        "doctrine/orm": "^2.6",
-        "symfony/validator": "^4.3",
-        "guzzlehttp/guzzle": "~6.0",
-        "laminas/laminas-eventmanager": "^3.10.0",
-        "composer/package-versions-deprecated": "^1.10.99",
-        "laminas/laminas-component-installer": "^2.6.0",
-        "laminas/laminas-config-aggregator": "^1.6",
-        "laminas/laminas-diactoros": "^2.7",
-        "laminas/laminas-stdlib": "^3.6",
-        "mezzio/mezzio": "^3.7",
-        "mezzio/mezzio-helpers": "^5.7",
-        "laminas/laminas-servicemanager": "^3.4",
-        "mezzio/mezzio-fastroute": "^3.0.3",
-        "jms/serializer": "^3.0",
         "ext-curl": "*",
         "ext-dom": "*",
         "ext-json": "*",
-        "ext-pdo": "*"
+        "ext-pdo": "*",
+        "composer/package-versions-deprecated": "^1.10.99",
+        "doctrine/annotations": "^1.6",
+        "doctrine/dbal": "^2.13.7 || ^3.3.2",
+        "doctrine/doctrine-laminas-hydrator": "^3.0.0",
+        "doctrine/doctrine-module": "^5.3.0 || ^6.0.2",
+        "doctrine/doctrine-orm-module": "^6.1",
+        "doctrine/event-manager": "^1.1.1",
+        "doctrine/migrations": "^3.3.0",
+        "doctrine/orm": "^2.11.1",
+        "doctrine/persistence": "^2.3.0 || ^3.0.0",
+        "guzzlehttp/guzzle": "~6.0",
+        "jms/serializer": "^3.0",
+        "laminas/laminas-cli": "^1.0.0",
+        "laminas/laminas-component-installer": "^2.6.0",
+        "laminas/laminas-config-aggregator": "^1.6",
+        "laminas/laminas-diactoros": "^2.7",
+        "laminas/laminas-eventmanager": "^3.10.0",
+        "laminas/laminas-modulemanager": "^2.11.0",
+        "laminas/laminas-mvc": "^3.3.2",
+        "laminas/laminas-paginator": "^2.12.2",
+        "laminas/laminas-servicemanager": "~3.3 || ~4.0",
+        "laminas/laminas-stdlib": "^3.7.1",
+        "mezzio/mezzio": "^3.7",
+        "mezzio/mezzio-authentication-oauth2": "^2.1",
+        "mezzio/mezzio-fastroute": "^3.0.3",
+        "mezzio/mezzio-helpers": "^5.7",
+        "psr/container": "^1.1.2",
+        "symfony/console": "^5.4.3 || ^6.0.3",
+        "symfony/validator": "^4.3"
     },
     "require-dev": {
         "laminas/laminas-development-mode": "^3.3.0",
