@@ -4,7 +4,26 @@
 namespace App\Repository;
 
 
-class CarrierRepository
-{
+use App\Entity\Carrier;
+use Doctrine\ORM\EntityRepository;
+use Throwable;
 
+class CarrierRepository extends EntityRepository
+{
+    /**
+     * @param $params
+     * @return int|array|string
+     * @throws Throwable
+     */
+    public function search($params): array|int|string
+    {
+        try {
+            $qb = $this->getEntityManager()->createQueryBuilder();
+            $qb->select('c')
+                ->from(Carrier::class, 'c');
+            return $qb->getQuery()->getArrayResult();
+        } catch (Throwable $e) {
+            throw $e;
+        }
+    }
 }
